# docs/api/api-docs.json
{"openapi":"3.1.0","info":{"title":"Findex API","description":"가볍고 빠른 외부 API 연동 금융 분석 도구 API 문서","license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"v1.0.0"},"servers":[{"url":"http://sprint-project-1196140422.ap-northeast-2.elb.amazonaws.com/sb/findex","description":"Generated server url"}],"tags":[{"name":"지수 정보 API","description":"지수 정보 관리 API"},{"name":"지수 데이터 API","description":"지수 데이터 관리 API"},{"name":"연동 작업 API","description":"연동 작업 관리 API"},{"name":"자동 연동 설정 API","description":"자동 연동 설정 관리 API"}],"paths":{"/api/sync-jobs/index-infos":{"post":{"tags":["연동 작업 API"],"summary":"지수 정보 연동","description":"Open API를 통해 지수 정보를 연동합니다.","operationId":"syncIndexInfo","responses":{"400":{"description":"잘못된 요청","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"202":{"description":"연동 작업 생성 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SyncJobDto"}}}}}}}},"/api/sync-jobs/index-data":{"post":{"tags":["연동 작업 API"],"summary":"지수 데이터 연동","description":"Open API를 통해 지수 데이터를 연동합니다.","operationId":"syncIndexData","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexDataSyncRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 날짜 범위 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"지수 정보를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"202":{"description":"연동 작업 생성 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SyncJobDto"}}}}}}}},"/api/index-infos":{"get":{"tags":["지수 정보 API"],"summary":"지수 정보 목록 조회","description":"지수 정보 목록을 조회합니다. 필터링, 정렬, 커서 기반 페이지네이션을 지원합니다.","operationId":"getIndexInfoList","parameters":[{"name":"indexClassification","in":"query","description":"지수 분류명","required":false,"schema":{"type":"string"}},{"name":"indexName","in":"query","description":"지수명","required":false,"schema":{"type":"string"}},{"name":"favorite","in":"query","description":"즐겨찾기 여부","required":false,"schema":{"type":"boolean"}},{"name":"idAfter","in":"query","description":"이전 페이지 마지막 요소 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","description":"커서 (다음 페이지 시작점)","required":false,"schema":{"type":"string"}},{"name":"sortField","in":"query","description":"정렬 필드 (indexClassification, indexName, employedItemsCount)","required":false,"schema":{"type":"string","default":"indexClassification"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (asc, desc)","required":false,"schema":{"type":"string","default":"asc"}},{"name":"size","in":"query","description":"페이지 크기","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 필터 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"지수 정보 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseIndexInfoDto"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"post":{"tags":["지수 정보 API"],"summary":"지수 정보 등록","description":"새로운 지수 정보를 등록합니다.","operationId":"createIndexInfo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexInfoCreateRequest"}}},"required":true},"responses":{"201":{"description":"지수 정보 생성 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndexInfoDto"}}}},"400":{"description":"잘못된 요청 (필수 필드 누락 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/index-data":{"get":{"tags":["지수 데이터 API"],"summary":"지수 데이터 목록 조회","description":"지수 데이터 목록을 조회합니다. 필터링, 정렬, 커서 기반 페이지네이션을 지원합니다.","operationId":"getIndexDataList","parameters":[{"name":"indexInfoId","in":"query","description":"지수 정보 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"startDate","in":"query","description":"시작 일자","required":false,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","description":"종료 일자","required":false,"schema":{"type":"string","format":"date"}},{"name":"idAfter","in":"query","description":"이전 페이지 마지막 요소 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","description":"커서 (다음 페이지 시작점)","required":false,"schema":{"type":"string"}},{"name":"sortField","in":"query","description":"정렬 필드 (baseDate, marketPrice, closingPrice, highPrice, lowPrice, versus, fluctuationRate, tradingQuantity, tradingPrice, marketTotalAmount)","required":false,"schema":{"type":"string","default":"baseDate"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (asc, desc)","required":false,"schema":{"type":"string","default":"desc"}},{"name":"size","in":"query","description":"페이지 크기","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 필터 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"지수 데이터 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseIndexDataDto"}}}}}},"post":{"tags":["지수 데이터 API"],"summary":"지수 데이터 등록","description":"새로운 지수 데이터를 등록합니다.","operationId":"createIndexData","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexDataCreateRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 데이터 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"201":{"description":"지수 데이터 생성 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndexDataDto"}}}},"404":{"description":"참조하는 지수 정보를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/index-infos/{id}":{"get":{"tags":["지수 정보 API"],"summary":"지수 정보 조회","description":"ID로 지수 정보를 조회합니다.","operationId":"getIndexInfo","parameters":[{"name":"id","in":"path","description":"지수 정보 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"지수 정보 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndexInfoDto"}}}},"404":{"description":"조회할 지수 정보를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["지수 정보 API"],"summary":"지수 정보 삭제","description":"지수 정보를 삭제합니다. 관련된 지수 데이터도 함께 삭제됩니다.","operationId":"deleteIndexInfo","parameters":[{"name":"id","in":"path","description":"지수 정보 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"삭제할 지수 정보를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"204":{"description":"지수 정보 삭제 성공"}}},"patch":{"tags":["지수 정보 API"],"summary":"지수 정보 수정","description":"기존 지수 정보를 수정합니다.","operationId":"updateIndexInfo","parameters":[{"name":"id","in":"path","description":"지수 정보 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexInfoUpdateRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 필드 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"수정할 지수 정보를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"지수 정보 수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndexInfoDto"}}}}}}},"/api/index-data/{id}":{"delete":{"tags":["지수 데이터 API"],"summary":"지수 데이터 삭제","description":"지수 데이터를 삭제합니다.","operationId":"deleteIndexData","parameters":[{"name":"id","in":"path","description":"지수 데이터 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"삭제할 지수 데이터를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"204":{"description":"지수 데이터 삭제 성공"},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"patch":{"tags":["지수 데이터 API"],"summary":"지수 데이터 수정","description":"기존 지수 데이터를 수정합니다.","operationId":"updateIndexData","parameters":[{"name":"id","in":"path","description":"지수 데이터 ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndexDataUpdateRequest"}}},"required":true},"responses":{"404":{"description":"수정할 지수 데이터를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 (유효하지 않은 데이터 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"지수 데이터 수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndexDataDto"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/auto-sync-configs/{id}":{"patch":{"tags":["자동 연동 설정 API"],"summary":"자동 연동 설정 수정","description":"기존 자동 연동 설정을 수정합니다.","operationId":"updateAutoSyncConfig","parameters":[{"name":"id","in":"path","description":"자동 연동 설정 ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AutoSyncConfigUpdateRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 설정 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"수정할 자동 연동 설정을 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"자동 연동 설정 수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AutoSyncConfigDto"}}}}}}},"/api/sync-jobs":{"get":{"tags":["연동 작업 API"],"summary":"연동 작업 목록 조회","description":"연동 작업 목록을 조회합니다. 필터링, 정렬, 커서 기반 페이지네이션을 지원합니다.","operationId":"getSyncJobList","parameters":[{"name":"jobType","in":"query","description":"연동 작업 유형 (INDEX_INFO, INDEX_DATA)","required":false,"schema":{"type":"string","description":"연동 작업 유형","enum":["INDEX_INFO","INDEX_DATA"]}},{"name":"indexInfoId","in":"query","description":"지수 정보 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"baseDateFrom","in":"query","description":"대상 날짜 (부터)","required":false,"schema":{"type":"string","format":"date"}},{"name":"baseDateTo","in":"query","description":"대상 날짜 (까지)","required":false,"schema":{"type":"string","format":"date"}},{"name":"worker","in":"query","description":"작업자","required":false,"schema":{"type":"string"}},{"name":"jobTimeFrom","in":"query","description":"작업 일시 (부터)","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"jobTimeTo","in":"query","description":"작업 일시 (까지)","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"status","in":"query","description":"작업 상태 (SUCCESS, FAILED)","required":false,"schema":{"type":"string","description":"연동 작업 상태","enum":["SUCCESS","FAILED"]}},{"name":"idAfter","in":"query","description":"이전 페이지 마지막 요소 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"cursor","in":"query","description":"커서 (다음 페이지 시작점)","required":false,"schema":{"type":"string"}},{"name":"sortField","in":"query","description":"정렬 필드 (targetDate, jobTime)","required":false,"schema":{"type":"string","default":"jobTime"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (asc, desc)","required":false,"schema":{"type":"string","default":"desc"}},{"name":"size","in":"query","description":"페이지 크기","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 필터 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"연동 작업 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseSyncJobDto"}}}}}}},"/api/index-infos/summaries":{"get":{"tags":["지수 정보 API"],"summary":"지수 정보 요약 목록 조회","description":"지수 ID, 분류, 이름만 포함한 전체 지수 목록을 조회합니다.","operationId":"getIndexInfoSummaries","responses":{"200":{"description":"지수 정보 요약 목록 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IndexInfoSummaryDto"}}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/index-data/{id}/chart":{"get":{"tags":["지수 데이터 API"],"summary":"지수 차트 조회","description":"지수의 차트 데이터를 조회합니다.","operationId":"getIndexChart","parameters":[{"name":"id","in":"path","description":"지수 정보 ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"periodType","in":"query","description":"차트 기간 유형 (DAILY, WEEKLY, MONTHLY)","required":false,"schema":{"type":"string","default":"DAILY","description":"차트 기간 유형","enum":["MONTHLY","QUARTERLY","YEARLY"]}}],"responses":{"200":{"description":"차트 데이터 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndexChartDto"}}}},"404":{"description":"지수 정보를 찾을 수 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 (유효하지 않은 기간 유형 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/index-data/performance/rank":{"get":{"tags":["지수 데이터 API"],"summary":"지수 성과 랭킹 조회","description":"지수의 성과 분석 랭킹을 조회합니다.","operationId":"getIndexPerformanceRank","parameters":[{"name":"indexInfoId","in":"query","description":"지수 정보 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"periodType","in":"query","description":"성과 기간 유형 (DAILY, WEEKLY, MONTHLY)","required":false,"schema":{"type":"string","default":"DAILY","description":"단위 기간 유형","enum":["DAILY","WEEKLY","MONTHLY"]}},{"name":"limit","in":"query","description":"최대 랭킹 수","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"성과 랭킹 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RankedIndexPerformanceDto"}}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"400":{"description":"잘못된 요청 (유효하지 않은 기간 유형 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/index-data/performance/favorite":{"get":{"tags":["지수 데이터 API"],"summary":"관심 지수 성과 조회","description":"즐겨찾기로 등록된 지수들의 성과를 조회합니다.","operationId":"getFavoriteIndexPerformance","parameters":[{"name":"periodType","in":"query","description":"성과 기간 유형 (DAILY, WEEKLY, MONTHLY)","required":false,"schema":{"type":"string","default":"DAILY","description":"단위 기간 유형","enum":["DAILY","WEEKLY","MONTHLY"]}}],"responses":{"200":{"description":"관심 지수 성과 조회 성공","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IndexPerformanceDto"}}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/index-data/export/csv":{"get":{"tags":["지수 데이터 API"],"summary":"지수 데이터 CSV export","description":"지수 데이터를 CSV 파일로 export합니다.","operationId":"downloadIndexData","parameters":[{"name":"indexInfoId","in":"query","description":"지수 정보 ID","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"startDate","in":"query","description":"시작 일자","required":false,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","description":"종료 일자","required":false,"schema":{"type":"string","format":"date"}},{"name":"sortField","in":"query","description":"정렬 필드 (baseDate, marketPrice, closingPrice, highPrice, lowPrice, versus, fluctuationRate, tradingQuantity, tradingPrice, marketTotalAmount)","required":false,"schema":{"type":"string","default":"baseDate"}},{"name":"sortDirection","in":"query","description":"정렬 방향 (asc, desc)","required":false,"schema":{"type":"string","default":"desc"}}],"responses":{"200":{"description":"CSV 파일 생성 성공","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}},"400":{"description":"잘못된 요청 (유효하지 않은 필터 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/auto-sync-configs":{"get":{"tags":["자동 연동 설정 API"],"summary":"자동 연동 설정 목록 조회","description":"자동 연동 설정 목록을 조회합니다. 필터링, ID 기준 커서 기반 페이지네이션을 지원합니다.","operationId":"getAutoSyncConfigList","parameters":[{"name":"idAfter","in":"query","description":"마지막 조회 ID (커서)","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"cursor","in":"query","description":"커서 UUID","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"indexInfoId","in":"query","description":"지수 정보 ID 필터","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"enabled","in":"query","description":"활성화 여부 필터","required":false,"schema":{"type":"boolean"}},{"name":"size","in":"query","description":"페이지 크기","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"400":{"description":"잘못된 요청 (유효하지 않은 필터 값 등)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"자동 연동 설정 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CursorPageResponseAutoSyncConfigDto"}}}},"500":{"description":"서버 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"ErrorResponse":{"type":"object","description":"에러 응답","properties":{"timestamp":{"type":"string","format":"date-time","description":"에러 발생 시간","example":"2025-03-06T05:39:06.152068Z"},"status":{"type":"integer","format":"int32","description":"HTTP 상태 코드","example":400},"message":{"type":"string","description":"에러 메시지","example":"잘못된 요청입니다."},"details":{"type":"string","description":"에러 상세 내용","example":"부서 코드는 필수입니다."}}},"SyncJobDto":{"type":"object","description":"연동 작업 DTO","properties":{"id":{"type":"integer","format":"int64","description":"연동 작업 ID","example":1},"jobType":{"type":"string","description":"연동 작업 유형","enum":["INDEX_INFO","INDEX_DATA"],"example":"INDEX_DATA"},"indexInfoId":{"type":"integer","format":"int64","description":"지수 정보 ID","example":1},"targetDate":{"type":"string","format":"date","description":"대상 날짜","example":"2023-01-01"},"worker":{"type":"string","description":"작업자","example":"192.168.0.1"},"jobTime":{"type":"string","format":"date-time","description":"작업 시간","example":"2023-01-01T12:00:00"},"result":{"type":"string","description":"작업 결과","enum":["SUCCESS","FAILED"],"example":"NEW"}}},"IndexDataSyncRequest":{"type":"object","description":"지수 데이터 연동 요청","properties":{"indexInfoIds":{"type":"array","description":"지수 정보 ID 목록 (비어있을 경우 모든 지수 대상)","example":[1,2,3],"items":{"type":"integer","format":"int64"}},"baseDateFrom":{"type":"string","format":"date","description":"대상 날짜 (부터)","example":"2023-01-01"},"baseDateTo":{"type":"string","format":"date","description":"대상 날짜 (까지)","example":"2023-01-31"}},"required":["baseDateFrom","baseDateTo"]},"IndexInfoCreateRequest":{"type":"object","description":"지수 정보 생성 요청","properties":{"indexClassification":{"type":"string","description":"지수 분류명","example":"KOSPI시리즈"},"indexName":{"type":"string","description":"지수명","example":"IT 서비스"},"employedItemsCount":{"type":"integer","format":"int32","description":"채용 종목 수","example":200},"basePointInTime":{"type":"string","format":"date","description":"기준 시점","example":"2000-01-01"},"baseIndex":{"type":"number","description":"기준 지수","example":1000.0},"favorite":{"type":"boolean","description":"즐겨찾기 여부","example":false}},"required":["baseIndex","basePointInTime","employedItemsCount","indexClassification","indexName"]},"IndexInfoDto":{"type":"object","description":"지수 정보 DTO","properties":{"id":{"type":"integer","format":"int64","description":"지수 정보 ID","example":1},"indexClassification":{"type":"string","description":"지수 분류명","example":"KOSPI시리즈"},"indexName":{"type":"string","description":"지수명","example":"IT 서비스"},"employedItemsCount":{"type":"integer","format":"int32","description":"채용 종목 수","example":200},"basePointInTime":{"type":"string","format":"date","description":"기준 시점","example":"2000-01-01"},"baseIndex":{"type":"number","description":"기준 지수","example":1000.0},"sourceType":{"type":"string","description":"출처 (사용자, Open API)","enum":["USER","OPEN_API"],"example":"OPEN_API"},"favorite":{"type":"boolean","description":"즐겨찾기 여부","example":true}}},"IndexDataCreateRequest":{"type":"object","description":"지수 데이터 생성 요청","properties":{"indexInfoId":{"type":"integer","format":"int64","description":"지수 정보 ID","example":1},"baseDate":{"type":"string","format":"date","description":"기준 일자","example":"2023-01-01"},"marketPrice":{"type":"number","description":"시가","example":2800.25},"closingPrice":{"type":"number","description":"종가","example":2850.75},"highPrice":{"type":"number","description":"고가","example":2870.5},"lowPrice":{"type":"number","description":"저가","example":2795.3},"versus":{"type":"number","description":"전일 대비 등락","example":50.5},"fluctuationRate":{"type":"number","description":"전일 대비 등락률","example":1.8},"tradingQuantity":{"type":"integer","format":"int64","description":"거래량","example":1250000},"tradingPrice":{"type":"integer","format":"int64","description":"거래대금","example":3500000000},"marketTotalAmount":{"type":"integer","format":"int64","description":"상장 시가 총액","example":450000000000}},"required":["baseDate","closingPrice","fluctuationRate","highPrice","indexInfoId","lowPrice","marketPrice","marketTotalAmount","tradingPrice","tradingQuantity","versus"]},"IndexDataDto":{"type":"object","description":"지수 데이터 DTO","properties":{"id":{"type":"integer","format":"int64","description":"지수 데이터 ID","example":1},"indexInfoId":{"type":"integer","format":"int64","description":"지수 정보 ID","example":1},"baseDate":{"type":"string","format":"date","description":"기준 일자","example":"2023-01-01"},"sourceType":{"type":"string","description":"출처 (사용자, Open API)","enum":["USER","OPEN_API"],"example":"OPEN_API"},"marketPrice":{"type":"number","description":"시가","example":2800.25},"closingPrice":{"type":"number","description":"종가","example":2850.75},"highPrice":{"type":"number","description":"고가","example":2870.5},"lowPrice":{"type":"number","description":"저가","example":2795.3},"versus":{"type":"number","description":"전일 대비 등락","example":50.5},"fluctuationRate":{"type":"number","description":"전일 대비 등락률","example":1.8},"tradingQuantity":{"type":"integer","format":"int64","description":"거래량","example":1250000},"tradingPrice":{"type":"integer","format":"int64","description":"거래대금","example":3500000000},"marketTotalAmount":{"type":"integer","format":"int64","description":"상장 시가 총액","example":450000000000}}},"IndexInfoUpdateRequest":{"type":"object","description":"지수 정보 수정 요청","properties":{"employedItemsCount":{"type":"integer","format":"int32","description":"채용 종목 수","example":200},"basePointInTime":{"type":"string","format":"date","description":"기준 시점","example":"2000-01-01"},"baseIndex":{"type":"number","description":"기준 지수","example":1000.0},"favorite":{"type":"boolean","description":"즐겨찾기 여부","example":true}}},"IndexDataUpdateRequest":{"type":"object","description":"지수 데이터 수정 요청","properties":{"marketPrice":{"type":"number","description":"시가","example":2800.25},"closingPrice":{"type":"number","description":"종가","example":2850.75},"highPrice":{"type":"number","description":"고가","example":2870.5},"lowPrice":{"type":"number","description":"저가","example":2795.3},"versus":{"type":"number","description":"전일 대비 등락","example":50.5},"fluctuationRate":{"type":"number","description":"전일 대비 등락률","example":1.8},"tradingQuantity":{"type":"integer","format":"int64","description":"거래량","example":1250000},"tradingPrice":{"type":"integer","format":"int64","description":"거래대금","example":3500000000},"marketTotalAmount":{"type":"integer","format":"int64","description":"상장 시가 총액","example":450000000000}}},"AutoSyncConfigUpdateRequest":{"type":"object","description":"자동 연동 설정 수정 요청","properties":{"enabled":{"type":"boolean","description":"활성화 여부","example":true}},"required":["enabled"]},"AutoSyncConfigDto":{"type":"object","description":"자동 연동 설정 DTO","properties":{"id":{"type":"string","format":"uuid","description":"자동 연동 설정 ID","example":"550e8400-e29b-41d4-a716-446655440000"},"indexInfoId":{"type":"string","format":"uuid","description":"지수 정보 ID","example":"550e8400-e29b-41d4-a716-446655440001"},"indexClassification":{"type":"string","description":"지수 분류명","example":"KOSPI시리즈"},"indexName":{"type":"string","description":"지수명","example":"IT 서비스"},"enabled":{"type":"boolean","description":"활성화 여부","example":true}}},"CursorPageResponseSyncJobDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/SyncJobDto"}},"nextCursor":{"type":"string","description":"다음 페이지 커서","example":"eyJpZCI6MjB9"},"nextIdAfter":{"type":"integer","format":"int64","description":"마지막 요소의 ID","example":"eyJpZCI6MjB9"},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}},"CursorPageResponseIndexInfoDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/IndexInfoDto"}},"nextCursor":{"type":"string","description":"다음 페이지 커서","example":"eyJpZCI6MjB9"},"nextIdAfter":{"type":"integer","format":"int64","description":"마지막 요소의 ID","example":"eyJpZCI6MjB9"},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}},"IndexInfoSummaryDto":{"type":"object","description":"지수 정보 요약 DTO","properties":{"id":{"type":"integer","format":"int64","description":"지수 정보 ID","example":1},"indexClassification":{"type":"string","description":"지수 분류명","example":"KOSPI시리즈"},"indexName":{"type":"string","description":"지수명","example":"IT 서비스"}}},"CursorPageResponseIndexDataDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/IndexDataDto"}},"nextCursor":{"type":"string","description":"다음 페이지 커서","example":"eyJpZCI6MjB9"},"nextIdAfter":{"type":"integer","format":"int64","description":"마지막 요소의 ID","example":"eyJpZCI6MjB9"},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}},"ChartDataPoint":{"type":"object","description":"차트 데이터 포인트","properties":{"date":{"type":"string","format":"date","description":"날짜","example":"2023-01-01"},"value":{"type":"number","description":"값","example":2850.75}}},"IndexChartDto":{"type":"object","description":"지수 차트 데이터 DTO","properties":{"indexInfoId":{"type":"integer","format":"int64","description":"지수 정보 ID","example":1},"indexClassification":{"type":"string","description":"지수 분류명","example":"KOSPI시리즈"},"indexName":{"type":"string","description":"지수명","example":"IT 서비스"},"periodType":{"type":"string","description":"차트 기간 유형 (DAILY, WEEKLY, MONTHLY)","enum":["MONTHLY","QUARTERLY","YEARLY"],"example":"DAILY"},"dataPoints":{"type":"array","description":"차트 데이터 목록","items":{"$ref":"#/components/schemas/ChartDataPoint"}},"ma5DataPoints":{"type":"array","description":"5일 이동평균선 데이터 목록","items":{"$ref":"#/components/schemas/ChartDataPoint"}},"ma20DataPoints":{"type":"array","description":"20일 이동평균선 데이터 목록","items":{"$ref":"#/components/schemas/ChartDataPoint"}}}},"IndexPerformanceDto":{"type":"object","description":"지수 성과 정보 DTO","properties":{"indexInfoId":{"type":"integer","format":"int64","description":"지수 정보 ID","example":1},"indexClassification":{"type":"string","description":"지수 분류명","example":"KOSPI시리즈"},"indexName":{"type":"string","description":"지수명","example":"IT 서비스"},"versus":{"type":"number","description":"단위 기간 대비 등락","example":50.5},"fluctuationRate":{"type":"number","description":"단위 기간 대비 등락률","example":1.8},"currentPrice":{"type":"number","description":"현재가","example":2850.75},"beforePrice":{"type":"number","description":"단위 기간 전 값","example":2850.75}}},"RankedIndexPerformanceDto":{"type":"object","description":"순위가 포함된 지수 성과 정보 DTO","properties":{"performance":{"$ref":"#/components/schemas/IndexPerformanceDto","description":"지수 성과 정보"},"rank":{"type":"integer","format":"int32","description":"순위","example":1}}},"CursorPageResponseAutoSyncConfigDto":{"type":"object","description":"커서 기반 페이지 응답","properties":{"content":{"type":"array","description":"페이지 내용","items":{"$ref":"#/components/schemas/AutoSyncConfigDto"}},"nextCursor":{"type":"string","format":"uuid","description":"다음 페이지 커서","example":"550e8400-e29b-41d4-a716-446655440000"},"nextIdAfter":{"type":"string","format":"uuid","description":"마지막 요소의 ID","example":"550e8400-e29b-41d4-a716-446655440000"},"size":{"type":"integer","format":"int32","description":"페이지 크기","example":10},"totalElements":{"type":"integer","format":"int64","description":"총 요소 수","example":100},"hasNext":{"type":"boolean","description":"다음 페이지 여부","example":true}}}}}}